Created packages for classes and moved all classes into relevant packages. Also removed expected exception test from NumberUtilsTest.
