+++ src/GameElements/Player.java
@@ -1,3 +1,5 @@
+package GameElements;
+
 public class Player {
 
     //Initialise variables for the player
